Move SAGESharp.IO.Trees namespace to the new project
