+++ SAGESharp.Tests/IO/Trees/TreeWithHeight1.cs
@@ -3,9 +3,10 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
+using SAGESharp.IO;
 using System.Collections.Generic;
 
-namespace SAGESharp.IO.Trees
+namespace SAGESharp.Tests.IO.Trees
 {
     static class TreeWithHeight1
     {
