+++ SAGESharp.Tests/IO/Trees/TreeWithHeight3.cs
@@ -3,11 +3,12 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
+using SAGESharp.IO;
 using System.Collections.Generic;
 
-namespace SAGESharp.IO.Trees
+namespace SAGESharp.Tests.IO.Trees
 {
-    static class TreeWithHeight2
+    static class TreeWithHeight3
     {
         public static IDataNode Build() => new BuilderFor.DataNodeSubstitute
         {
@@ -15,24 +16,24 @@ namespace SAGESharp.IO.Trees
             {
                 new BuilderFor.EdgeSubstitute<Class>
                 {
-                    ChildNode = new BuilderFor.DataNodeSubstitute().Build(),
-                    ChildExtractor = value => value.Long,
-                    ChildSetter = (value, childValue) => value.Long = (long)childValue
+                    ChildNode = TreeWithHeight2.Build(),
+                    ChildExtractor = value => value.Child1,
+                    ChildSetter = (value, childValue) => value.Child1 = (TreeWithHeight2.Class)childValue
                 }.Build(),
                 new BuilderFor.EdgeSubstitute<Class>
                 {
-                    ChildNode = TreeWithHeight1.Build(),
-                    ChildExtractor = value => value.Child,
-                    ChildSetter = (value, childValue) => value.Child = (TreeWithHeight1.Class)childValue
+                    ChildNode = TreeWithHeight2.Build(),
+                    ChildExtractor = value => value.Child2,
+                    ChildSetter = (value, childValue) => value.Child2 = (TreeWithHeight2.Class)childValue
                 }.Build()
             }
         }.Build();
 
         public class Class
         {
-            public long Long { get; set; }
+            public TreeWithHeight2.Class Child1 { get; set; }
 
-            public TreeWithHeight1.Class Child { get; set; }
+            public TreeWithHeight2.Class Child2 { get; set; }
         }
     }
 }
